procureflow_env/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "procureflow-env"
version = "0.1.0"
description = "OpenEnv-compatible procurement and vendor operations simulation environment."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "openai>=1.0.0,<2.0.0",
  "openenv-core>=0.2.1",
  "pydantic>=2.8.0,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "pytest>=8.0.0,<9.0.0",
  "requests>=2.31.0,<3.0.0",
  "uvicorn>=0.30.0,<1.0.0",
]

[project.scripts]
server = "app.server:main"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[tool.setuptools]
include-package-data = true
packages = [
  "app",
  "baseline",
  "graders",
  "server",
  "tasks",
]
py-modules = ["client", "inference", "models"]
